15/11/22

+++ src/ntcc/Tasks.java
@@ -2,12 +2,12 @@ package ntcc;
 
 public class Tasks {
 
-    int burst_time=0;
+    int burst_time = 0;
 
-    int arrival_time=0;
-    int deadline=0;
+    int arrival_time = 0;
+    int deadline = 0;
 
-    int waiting_time=0;
-    float membership_type=0.0f;
+    int waiting_time = 0;
+    double membership_type = 0.0f;
 
 }
